requirement.txt
streamlit
pandas
pydeck
plotly.graph_objs
